+++ tests/oihana/reflect/mocks/MockMarker.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace tests\oihana\reflect\mocks;
+
+use Attribute;
+
+/**
+ * A multi-target attribute used to test the attribute readers.
+ */
+#[Attribute( Attribute::TARGET_CLASS | Attribute::TARGET_PROPERTY | Attribute::TARGET_METHOD )]
+class MockMarker
+{
+    public function __construct( public string $tag = '' )
+    {
+    }
+}